multi-gpu training

+++ models/sae.py
@@ -6,7 +6,6 @@ import wandb
 import tempfile
 import os
 
-
 class SparseAutoencoder(nn.Module):
     def __init__(self, hyperparameters: Dict[str, Any]):
         super().__init__()
@@ -41,12 +40,13 @@ class SparseAutoencoder(nn.Module):
         return x * (x >= top_values[:, -1:])
 
     def save_model(self, run_name: str, alias: str="latest"):
-        artifact = wandb.Artifact(run_name, type='model')
-        with tempfile.NamedTemporaryFile(mode='wb', delete=False, suffix='.pth') as tmp_file:
-            torch.save(self.state_dict(), tmp_file.name)
-            artifact.add_file(tmp_file.name, f'{run_name}.pth')
-        wandb.log_artifact(artifact, aliases=[alias])
-        os.remove(tmp_file.name)
+        if torch.distributed.get_rank() == 0:
+            artifact = wandb.Artifact(run_name, type='model')
+            with tempfile.NamedTemporaryFile(mode='wb', delete=False, suffix='.pth') as tmp_file:
+                torch.save(self.state_dict(), tmp_file.name)
+                artifact.add_file(tmp_file.name, f'{run_name}.pth')
+            wandb.log_artifact(artifact, aliases=[alias])
+            os.remove(tmp_file.name)
 
     @classmethod
     def load_from_pretrained(cls, artifact_path: str, hyperparameters, device="cpu"):
@@ -58,4 +58,3 @@ class SparseAutoencoder(nn.Module):
             model.load_state_dict(torch.load(model_path, map_location=device))
             model.to(device)
             return model
-
